Merged PR 140: Merge workaround for VSTS SOAP Query Parameters exception
It looks like there is a bug in VSTS that prevents the use of @ parameters in queries. This results in a SOAP error as reported in [VSTS] #5066 [GitHub] #103, #104. I added my own pre-processor for the query that replaces all of the @ parameters with the actual values. Integers are left open and strings are wrapped in '. I can see future problems, but this issue has been reported to MSFT and hopefully a fix will remove the need for this work around.

Related work items: #5066