Fixed bug: migrating from TFS to TFS project showed 'object reference not set to an instance of an object' error in log. Fixed it by using null-conditional operators. (#111)
The code was checking if a string (reflected id field value) was null. ToString() was used to convert field value (Value property of field) to string. If the field was null, a 'object reference not set to an instance of an object' exception. An exception would end up in the log being meaningless to a user not displayed to the user. If the value was null, then the ToString() would cause an exception, being called on a null object.